requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
Pillow==5.4.1
redis==3.2.1
requests==2.21.0
rq==0.13.0
urllib3==1.24.1
Werkzeug==0.15.1
